# packages/contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: WitJoy AI Academy API
  version: 0.2.0
  description: Versioned API contract for 智趣AI学堂.
servers:
  - url: /api/v1
...
paths:
...
  /ai/tasks/questions:
    servers:
      - url: /api/v2
    post:
      tags: [AI Tasks]
      summary: Create an asynchronous AI question task
      operationId: createAiQuestionTask
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiQuestionGenerateRequest"
      responses:
        "202":
          description: Task accepted for asynchronous processing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiTaskCreateApiResponse"
        "401":
          description: Authentication required
  /ai/tasks/{taskId}:
    servers:
      - url: /api/v2
    get:
      tags: [AI Tasks]
      summary: Read an owned AI task status
      operationId: getAiTask
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Current task status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiTaskStatusApiResponse"
        "401":
          description: Authentication required
        "404":
          description: Task not found for the authenticated user
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          const: 200
        message:
          type: string
        data: {}
    AiQuestionGenerateRequest:
      type: object
      required: [subject, difficulty]
      additionalProperties: false
      properties:
        subject:
          type: string
          minLength: 1
          example: 人工智能基础
        difficulty:
          type: string
          minLength: 1
          example: 基础
    AiQuestion:
      type: object
      required: [question, options, answer, analysis, difficulty]
      properties:
        question:
          type: string
        options:
          type: array
          minItems: 4
          maxItems: 4
          items:
            type: string
        answer:
          type: string
        analysis:
          type: string
        difficulty:
          type: string
    AiTaskCreate:
      type: object
      required: [taskId, status]
      properties:
        taskId:
          type: string
        status:
          type: string
          enum: [PENDING]
    AiTaskStatus:
      type: object
      required: [taskId, status, result]
      properties:
        taskId:
          type: string
        status:
          type: string
          enum: [PENDING, RUNNING, SUCCESS, REVIEW_REQUIRED, FAILED]
        result:
          type: [array, "null"]
          items:
            $ref: "#/components/schemas/AiQuestion"
        errorMessage:
          type: [string, "null"]
    AiTaskCreateApiResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/AiTaskCreate"
    AiTaskStatusApiResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/AiTaskStatus"